+++ __tests__/controllers/modules/auth.test.js
@@ -18,19 +18,19 @@ beforeEach(() => {
 // test
 describe('Modules: Auth', () => {
   test('Middleware: isSignedIn', () => {
-    const isSignedInCore = isSignedIn[isSignedIn.length - 1];
+    const _isSignedIn = isSignedIn[isSignedIn.length - 1];
     req.session.user = {};
 
     // should throw exception if authenticated
     req.isAuthenticated = () => false;
-    expect(() => isSignedInCore(req, res, next)).toThrowError(expect.objectContaining({
+    expect(() => _isSignedIn(req, res, next)).toThrowError(expect.objectContaining({
       name: 'ClientException',
       code: 20000,
     }));
 
     // should not throw exception if authenticated
     req.isAuthenticated = () => true;
-    expect(() => isSignedInCore(req, res, next)).not.toThrowError();
+    expect(() => _isSignedIn(req, res, next)).not.toThrowError();
 
     // should call `next` only once
     expect(next).toHaveBeenCalledTimes(1);
@@ -38,13 +38,13 @@ describe('Modules: Auth', () => {
 
 
   test('Middleware: isAuthorized', async () => {
-    const isAuthorizedCore = isAuthorized[isAuthorized.length - 1];
+    const _isAuthorized = isAuthorized[isAuthorized.length - 1];
     req.user = { _id: 'some_user_id' };
 
     // should throw exception if not authorized
     try {
       PostsModel.count = jest.fn(() => 2);
-      await isAuthorizedCore(req, res, next);
+      await _isAuthorized(req, res, next);
     } catch (err) {
       expect(err).toEqual(expect.objectContaining({
         name: 'ClientException',
@@ -54,7 +54,7 @@ describe('Modules: Auth', () => {
 
     // should NOT throw exception if authorized
     PostsModel.count = jest.fn(() => 1);
-    await isAuthorizedCore(req, res, next);
+    await _isAuthorized(req, res, next);
 
     // should call `next` only once
     expect(next).toHaveBeenCalledTimes(1);